requirements.txt
Django==3.2.7
djangorestframework==3.12.4
django-cors-headers==3.9.0
vue-color-picker-wheel==0.4.3
vue-router==3.5.2
vue-video-player==5.0.2
vue2-perfect-scrollbar==1.5.0
axios==0.22.0
drf-yasg==1.20.0
django-rest-swagger==2.2.0
